Refactor external share receiver logic from ShareRecorder into own component
